services/api/requirements.txt
django>=5.0,<5.1
djangorestframework>=3.15,<3.16
django-cors-headers>=4.4,<5.0
psycopg[binary]>=3.1,<4.0
python-dotenv>=1.0,<2.0
celery>=5.3,<6.0
redis>=5.0,<6.0
drf-spectacular>=0.27,<0.28
djangorestframework-simplejwt>=5.3,<6.0
reportlab>=4.1,<5.0
